[gunicorn] Use TypedDict for proxy_protocol_info
