+++ stubs/gunicorn/gunicorn/asgi/parser.pyi
@@ -1,6 +1,6 @@
 from collections.abc import Callable, Iterable
 from enum import IntEnum
-from typing import Any, Final, Literal, SupportsIndex, TypeAlias, TypedDict, type_check_only
+from typing import Any, Final, Literal, SupportsIndex, TypeAlias
 from typing_extensions import Self
 
 _H1CProtocol: TypeAlias = Any  # gunicorn_h1c H1CProtocol class
@@ -38,21 +38,7 @@ class UnsupportedTransferCoding(ParseError): ...
 class InvalidChunkSize(ParseError): ...
 class InvalidChunkExtension(ParseError): ...
 
-@type_check_only
-class _ProxyProtocolInfo(TypedDict):
-    proxy_protocol: Literal["TCP4", "TCP6", "UDP4", "UDP6"]
-    client_addr: str
-    client_port: int
-    proxy_addr: str
-    proxy_port: int
-
-@type_check_only
-class _ProxyProtocolInfoUnknown(TypedDict):
-    proxy_protocol: Literal["UNKNOWN", "LOCAL", "UNSPEC"]
-    client_addr: None
-    client_port: None
-    proxy_addr: None
-    proxy_port: None
+from .._types import _ProxyProtocolInfoDict
 
 class PythonProtocol:
     __slots__ = (
@@ -114,7 +100,7 @@ class PythonProtocol:
     ) -> None: ...
     def feed(self, data: Iterable[SupportsIndex]) -> None: ...
     @property
-    def proxy_protocol_info(self) -> _ProxyProtocolInfo | _ProxyProtocolInfoUnknown | None: ...
+    def proxy_protocol_info(self) -> _ProxyProtocolInfoDict | None: ...
     def reset(self) -> None: ...
     def finish(self) -> None: ...
 
@@ -149,7 +135,7 @@ class CallbackRequest:
     content_length: int
     chunked: bool
     must_close: bool
-    proxy_protocol_info: dict[str, str | int | None] | None  # TODO: Use TypedDict
+    proxy_protocol_info: _ProxyProtocolInfoDict | None
 
     def __init__(self) -> None: ...
     @classmethod
