+++ stubs/gunicorn/gunicorn/asgi/uwsgi.pyi
@@ -5,7 +5,7 @@ from gunicorn.asgi.unreader import AsyncUnreader
 from gunicorn.config import Config
 from gunicorn.uwsgi.message import UWSGIRequest
 
-from .._types import _AddressType
+from .._types import _AddressType, _ProxyProtocolInfoDict
 
 class AsyncUWSGIRequest(UWSGIRequest):
     cfg: Config
@@ -26,7 +26,7 @@ class AsyncUWSGIRequest(UWSGIRequest):
     uwsgi_vars: dict[str, str]
     modifier1: int
     modifier2: int
-    proxy_protocol_info: dict[str, str | int | None] | None  # TODO: Use TypedDict
+    proxy_protocol_info: _ProxyProtocolInfoDict | None
     content_length: int
     chunked: bool
