+++ stubs/gunicorn/gunicorn/http/message.pyi
@@ -7,7 +7,7 @@ from gunicorn.config import Config
 from gunicorn.http.body import Body
 from gunicorn.http.unreader import Unreader
 
-from .._types import _AddressType
+from .._types import _AddressType, _ProxyProtocolInfoDict
 
 PP_V2_SIGNATURE: Final = b"\x0d\x0a\x0d\x0a\x00\x0d\x0a\x51\x55\x49\x54\x0a"
 
@@ -66,7 +66,7 @@ class Request(Message):
     fragment: str | None
     limit_request_line: int
     req_number: int
-    proxy_protocol_info: dict[str, str | int | None] | None  # TODO: Use TypedDict
+    proxy_protocol_info: _ProxyProtocolInfoDict | None
 
     def __init__(self, cfg: Config, unreader: Unreader, peer_addr: _AddressType, req_number: int = 1) -> None: ...
     def get_data(self, unreader: Unreader, buf: io.BytesIO, stop: bool = False) -> None: ...
