+++ stubs/gunicorn/gunicorn/uwsgi/message.pyi
@@ -4,7 +4,7 @@ from gunicorn.config import Config
 from gunicorn.http.body import Body
 from gunicorn.http.unreader import Unreader
 
-from .._types import _AddressType
+from .._types import _AddressType, _ProxyProtocolInfoDict
 
 MAX_UWSGI_VARS: Final = 1000
 
@@ -28,7 +28,7 @@ class UWSGIRequest:
     uwsgi_vars: dict[str, str]
     modifier1: int
     modifier2: int
-    proxy_protocol_info: dict[str, str | int | None] | None  # TODO: Use TypedDict
+    proxy_protocol_info: _ProxyProtocolInfoDict | None
 
     def __init__(self, cfg: Config, unreader: Unreader, peer_addr: _AddressType, req_number: int = 1) -> None: ...
     def force_close(self) -> None: ...
